[edit] - synced tables / testing travis

+++ rest-server/src/config/database/models/dateCourse.js
@@ -28,7 +28,7 @@ module.exports = db.define('dateCourse', {
     allowNull: false
   },
 	rating: {
-    type: Sequelize.tinyint,
+    type: Sequelize.INTEGER,
     allowNull: false
   },
 	coments: {
@@ -36,11 +36,11 @@ module.exports = db.define('dateCourse', {
     allowNull: false
   },
 	creator: {
-    type: Sequelize.tinyint,
+    type: Sequelize.INTEGER,
     allowNull: false
   },
 	rating_count: {
-    type: Sequelize.tinyint,
+    type: Sequelize.INTEGER,
     allowNull: false
   },
